add controller placement function
This function contructs all the possible combinations
given the number of nodes and controllers.

While here modify the metrics functions to expect
controller placement dictionary of dictionaries
to be set with int(1) value only in the node that
exists a controller. All the other nodes will not
have a key into the dictionaries.

Signed-off-by: Tudor Ambarus <tudor.ambarus@gmail.com>