mcda, draw, algs: start counting controllers from zero
*mcda:
- start counting controllers from zero
Static execution had the Controller Placement dictionary
with key of value int(1). MCDA alg was written for this
dictionary and missed the possibility of a Controller
Placement that started with index int(0).

Update the static dictionary to begin with index int(0)

*algs:
- rename variables in average latency function

*draw:
- remove computation in print function

This patch will be squashed with the initial patch.

Signed-off-by: Claudiu Dumitru Iosub <i.claudiudumitru@gmail.com>